Pick values from mapping data

Pick values from mapping data based on coordinates of x and y.

pick(.data, x = NULL, y = NULL, phase = if (is.null(x)) NULL else
  paste0("P", seq_along(x)), i = data.frame(x, y, phase), ...)

# S3 method for qm_xmap
pick(.data, x = NULL, y = NULL, phase = if
  (is.null(x)) NULL else paste0("P", seq_along(x)), i = data.frame(x, y,
  phase), ...)

Arguments

.data
Mapping data. Currentl version only supports qm_xmap object.

x, y
Integer vectors specifying x- and y-coordinates of pixels of mapping data. In other words, x- and y- corresponds to n-th column and row of a matrix.

phase
Name of phases comprising the pixels to be picked. If not manually supplied phase is supplied by c('P1', 'P2', ...).

i
A data frame with clumns x, y, and phase. i have priority to x, y, and phase.`

...
Ignored

Value

data.frame

See also

Examples

# For qm_xmap class object
xm <- list(
  Si = as.data.frame(matrix(1:4, 2)),
  Ti = as.data.frame(matrix(5:8, 2))
)
class(xm) <- c('qm_xmap', 'list')

## following codes return same results
pick(xm, x = 1:2, y = 1:2)

#>   phase Si Ti
#> 1    P1  1  5
#> 2    P2  4  8

pick(xm, x = 1:2, y = 1:2, phase = c('P1', 'P2'))

#>   phase Si Ti
#> 1    P1  1  5
#> 2    P2  4  8

pick(xm, i = data.frame(x = 1:2, y = 1:2, phase = c('P1', 'P2')))

#>   phase Si Ti
#> 1    P1  1  5
#> 2    P2  4  8

## The last code is especially useful in case data frame is saved as a file such as csv.

# NOT RUN {
pick(xm, i = read.csv('pick_centers.csv'))
# }