Interactively quantify X-ray maps

By default, it starts browser to open web-based UI. All arguments are optional.

qntmap(
  xmap_dir = NULL,
  qnt_dir = NULL,
  deadtime = 1100,
  phase_list = NULL,
  console = FALSE,
  selected = c("auto", "Input", "Map", "Spot")
)

Arguments

xmap_dir, qnt_dir
Paths to the directories containing data from mapping and spot analysis.

deadtime
Dead time in nano seconds.

phase_list
A path to a csv file containing columns indicating phase of each analysis and TRUE or FALSE to use it for quantifying.

console
Wheter to use console-based interactive mode (default: FALSE). Notet hat the console-based one is retiring and no more maintained.

selected
A starting page of shiny UI. One of the "auto", "Input", "Map" or "Spot". If auto (default), "Map" is chosen when xmap_dir is specified, or "Spot" when qnt_dir is specified, or "Input" when both are not specified.