(Deprecated) Use quantify.

(Deprecated) Use quantify.

qntmap_quantify(wd = ".", dir_map, RDS_cluster, maps_x = NULL,
  maps_y = NULL, fine_phase = NULL, fine_th = 0.9,
  qnt = qnt_load(wd), qltmap = qltmap_load(dir_map),
  cluster = readRDS(RDS_cluster))

Arguments

wd
working directory which contains .qnt and .map directories

dir_map
ignored

RDS_cluster
path to the RDS file created by cluster_xmap

maps_x
A x-axis size of maps comprising guide net map (default: NULL).

maps_y
A y-axis size of maps comprising guide net map (default: NULL).

fine_phase
A character vector to specify fine grained phases which tend to comprise multi-phase pixels e.g., c('Pl', 'Amp')

fine_th
0.9

qnt
qm_qnt class object returned by read_qnt().

qltmap
qm_xmap class object

cluster
qm_cluster class object returned by cluter_xmap().