Quantify X-ray maps

Quantify X-ray maps

quantify(
  xmap,
  qnt,
  cluster,
  maps_x = attr(xmap, "pixel")[1L],
  maps_y = attr(xmap, "pixel")[2L],
  phase = everything(),
  element = everything(),
  fine_phase = NULL,
  fine_th = 0.9,
  fix = NULL,
  se = FALSE,
  saving = TRUE,
  ...
)

Arguments

xmap
Object returned by read_xmap().

qnt
Object returned by read_qnt().

cluster
Object returned by cluster_xmap().

maps_x, maps_y
Sizes of maps along x- and y-axes comprising guide net map. (default: NULL).

phase
Selected ones are referenced to detect outliers. Default selects everything. Tidy selection is available. For example c(Si, Ti) selects them, and c(-Si, -Ti) selects everything except them.

element
Selected ones are referenced to detect outliers. Default selects everything. Tidy selection is available. For example c(Si, Ti) selects them, and c(-Si, -Ti) selects everything except them.

fine_phase
Deprecated as of qntmap > 0.4.0. Use phase instead.

fine_th
A threshold of membership degrees to 0.9

fix
A path to the file specifying chemical compositions of some elements in some phases (default: NULL).

se
FALSE in default, and is forced when fix is specified. TRUE calculates standard errors, but require large memories.

saving
TRUE (default) saves the results into qntmap directory under the directory xmap is read from. FALSE does not save.`

...
Arguments passed on to find_outlier

interval

A type of the interval. Data points outside intervals are treated as outliers. If "prediction" (default), prediction intervals are used based on Poisson process. If `"tukey"``, conditional lower and upper whiskers are used where the lower is $Q_1/4 - 1.5 IQR$ and the upper is $Q_3/4 + 1.5 IQR$.

method

Applicable when interval = "prediction". If "rq" (default), quantile regression is performed (quantreg::rq()). If "lsfit", least square regression is performed (stats::lsfit()).

percentile

A percentile of predictive interaval. Applicable when interval = "prediction".