read .cnd files

read .cnd files

read_cnd(x, pattern = NULL, ...)

# S3 method for default
read_cnd(x, pattern = NULL, ...)

# S3 method for map_cnd
read_cnd(x, pattern = NULL, ...)

# S3 method for 0_cnd
read_cnd(x, pattern = NULL, n = NULL, ...)

Arguments

x
A path to the file (e.g., '0.cnd', 'map.cnd', ...)

pattern
pattern to substract

...
Other arguments passed to methods

n
integer vector of same length as pattern. Used to extract nth row of .cnd file in case pattern did not match any phrase.

.default

A default method which returns a result of readLines() with additional class according to the content of the file.

read_cnd.map_cnd

A method for map_cnd class object.

read_cnd.0_cnd

A method for 0_cnd class object. Used to extract rows which contains phrase matching pattern.