S3 generics to save returned values from functions in QntMap package.

S3 generics to save returned values from functions in QntMap package.

save4qm(x, nm = "", saving = TRUE, ...)

# S3 method for default
save4qm(x, nm, saving, ...)

# S3 method for data.frame
save4qm(x, nm, saving, ...)

# S3 method for qm_cluster
save4qm(x, nm, saving, ...)

# S3 method for qntmap
save4qm(x, nm, dir_qntmap, ...)

Arguments

x
object to save

nm
name of object to save

saving
TRUE or FALSE to save result file(s).

...
other arguments passed to class methods

dir_qntmap
A path to the directory to save x.

Value

invisible(x)

save4qm.default

A default method. This is a wrapper of saveRDS together with invisible return of x.

save4qm.data.frame

A method for data.frame class. This is a wrapper of data.table::fwrite together with invisible return of x.

save4qm.qm_cluster

A method for qm_cluster class returned by cluster_xmap() or cluster_group(). Object is saved as RDS file and png file. The latter shows distribution of phases among a mapped area.

save4qm.qntmap

A method for qntmap class returned by qntmap() or quantify()